pyproject.toml
[tool.poetry]
name = "creditcard"
version = "0.1.0"
description = ""
authors = ["GuiCastroo <gscastroo15@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.4.2"
cryptography = "^41.0.5"
fastapi = "^0.104.1"
sqlalchemy = "2.0.23"
uvicorn = "^0.23.2"
asyncpg = "^0.28.0"
fastapi-pagination = "^0.12.12"
httpx = "^0.25.1"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
pyjwt = "^2.8.0"
python-multipart = "^0.0.6"
alembic = "^1.12.1"
pydantic-settings = "^2.0.3"
python-dotenv = "^1.0.0"
greenlet = "^3.0.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
